# core/core-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "The Natixar ESG impact assessment open API"
  },
  "servers": [
    {
      "description": "Operational API",
      "url": "https://api.natixar.pro/core/{version}",
      "variables": {
        "version": {
          "enum": [
            "v1",
            "v2",
            "v3"
          ],
          "default": "v1"
        }
      }
    }
  ],
  "paths": {
...
    "/clients": {
      "post": {
        "summary": "Create a new client",
        "description": "Create a new client with the provided name and description",
        "operationId": "createClient",
        "tags": [
          "Clients"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "components-openapi.json#/components/schemas/Client"
              },
              "example": {
                "name": "Agro Novae Industrie, SAS",
                "id": 1
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a client. URN is returned in the Location header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "components-openapi.json#/components/schemas/Client"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "components-openapi.json#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "components-openapi.json#/components/responses/Unauthorized401"
          },
          "500": {
            "$ref": "components-openapi.json#/components/responses/InternalServerError500"
          }
        }
      },
      "get": {
        "summary": "List all clients",
        "description": "List all clients. The clients are returned in a list.",
        "operationId": "listClients",
        "tags": [
          "Clients"
        ],
        "responses": {
          "200": {
            "description": "A list of clients",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "components-openapi.json#/components/schemas/Client"
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "components-openapi.json#/components/responses/InternalServerError500"
          }
        }
      }
    },
    "/clients/{clientId}": {
      "get": {
        "summary": "Get a client by ID",
        "description": "Get a client by ID. If the client does not exist, a 404 error is returned.",
        "operationId": "getClient",
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A client object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "components-openapi.json#/components/schemas/Client"
                }
              }
            }
          },
          "404": {
            "$ref": "components-openapi.json#/components/responses/ClientNotFound404"
          },
          "500": {
            "$ref": "components-openapi.json#/components/responses/InternalServerError500"
          }
        }
      },
      "put": {
        "summary": "Update a client by ID",
        "description": "Update a client by ID. If the client does not exist, a 404 error is returned.",
        "operationId": "updateClient",
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "components-openapi.json#/components/schemas/Client"
              },
              "example": {
                "name": "Agro Novae Industrie, SAS",
                "id": 1
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the client",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "components-openapi.json#/components/schemas/Client"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input provided"
          },
          "404": {
            "$ref": "components-openapi.json#/components/responses/ClientNotFound404"
          },
          "500": {
            "$ref": "components-openapi.json#/components/responses/InternalServerError500"
          }
        }
      },
      "delete": {
        "summary": "Delete a client by ID",
        "description": "Delete a client by ID. If the client does not exist, a 404 error is returned.",
        "operationId": "deleteClient",
        "tags": [
          "Clients"
        ],
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted the client"
          },
          "404": {
            "description": "Client not found"
          },
          "500": {
            "$ref": "components-openapi.json#/components/responses/InternalServerError500"
          }
        }
      }
    },
...
    "/data/ranges": {
      "x-comment" : "This endpoint is deprecated, please use /scenarios/{id}/ranges",
      "get": {
        "summary": "Get a cube of data from the default scenario.",
        "operationId": "getData",
        "description": "### Function\nRetrieve the environmental impact data displayed by the dashboards. The data is discretized at the requested timescale and categorized according to the requested taxonomy.\n\n### Usage\nExample:\n```\nGET /core/v0/data/ranges?time_ranges=%5B%7B%22start%22%3A%222023-01-01T00%3A00%3A00+02%3A00%22%2C%22end%22%3A%222023-01-02T00%3A00%3A00+02%3A00%22%7D%2C%7B%22start%22%3A%222023-02-01T00%3A00%3A00+02%3A00%22%2C%22end%22%3A%222023-02-02T00%3A00%3A00+02%3A00%22%7D%5D&scale=1d&protocol=begesv5\nHost: api.natixar.pro\n\nuser-agent: curl/7.68.0\naccept: */*\n```\n### Returns\nThe endpoint returns exactly the same JSON object as /scenarios/{scenarioId}/ranges.\n\n### Remarks\nDeprecated. Superseded by /scenarios/{scenarioId}/ranges.",
        "externalDocs": {
          "description": "# API Specification-Data Endpoint (current revision)",
          "url": "https://docs.google.com/document/d/1kndLG5AxwOdeWPuOYKocnGJPY0TvwHj0zQ6aHXcIzdc"
        },
        "tags": [
          "Scenario",
          "Data",
          "Time Ranges",
          "Deprecated"
        ],
        "security": [
          {
            "client_api_key": []
          }
        ],
        "parameters": [
          {
            "name": "time_ranges",
            "in": "query",
            "description": "JSON array of time range objects, each specifying a start and an end time.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "json"
            },
            "example": "[{\"start\":\"2023-01-01T00:00:00+02:00\",\"end\":\"2023-01-02T00:00:00+02:00\"},{\"start\":\"2023-02-01T00:00:00+02:00\",\"end\":\"2023-02-02T00:00:00+02:00\"}]"
          },
          {
            "name": "scale",
            "in": "query",
            "description": "Specifies the scale of data aggregation.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "1d"
          },
          {
            "name": "protocol",
            "in": "query",
            "description": "Specifies the data protocol used.",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "begesv5"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "emissions-openapi.json#/components/schemas/EmissionsData"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "admin_api_key": {
        "type": "apiKey",
        "description": "Temporary security scheme based on an API key: back office API",
        "in": "header",
        "name": "x-admin-api-key"
      },
      "client_api_key": {
        "type": "apiKey",
        "description": "Temporary security scheme based on an API key: client API",
        "in": "header",
        "name": "x-client-api-key"
      }
    }
  },
  "security": [
    {
      "admin_api_key": []
    }
  ]
}